modify route structure
