+++ app/guid-node/metadata/add/route.ts
@@ -8,14 +8,13 @@ export default class GuidMetadataAddRoute extends Route {
     @service store!: Store;
 
     async model() {
-        const params = this.modelFor('guid-node.metadata');
-
+        const target = await this.modelFor('guid-node.metadata').taskInstance;
         const templates = await this.store.findAll('cedar-metadata-template', {
             adapterOptions: { sort: 'schema_name' },
         });
 
         return {
-            target: params.guidNode,
+            target,
             templates,
         };
     }
