+++ app/guid-node/metadata/detail/controller.ts
@@ -2,7 +2,7 @@ import Controller from '@ember/controller';
 import { inject as service } from '@ember/service';
 import Media from 'ember-responsive';
 
-export default class GuidMetadata extends Controller {
+export default class GuidMetadataDetail extends Controller {
     @service media!: Media;
