+++ app/guid-node/metadata/detail/styles.scss
