+++ app/guid-node/metadata/detail/template.hbs
@@ -23,7 +23,7 @@
                         data-test-add-cedar-schema
                         data-analytics-name='Add metadata schema'
                         @route='guid-node.metadata.add'
-                        @models={{array this.model.guidNode.id}}
+                        @models={{array this.model.target.id}}
                     >
                         {{t 'metadata.add-schema'}}
                     </OsfLink>
@@ -33,7 +33,7 @@
     </div>
     <div local-class='metadata-tab-container'>
         <GuidNode::Metadata::-Components::MetadataTabs
-            @guidNode={{this.model.guidNode}}
+            @guidNode={{this.model.target}}
             @cedarMetadataRecords={{this.model.cedarMetadataRecords}}
         ></GuidNode::Metadata::-Components::MetadataTabs>
     </div>
