+++ app/guid-node/metadata/index/route.ts
@@ -1,30 +1,8 @@
-import Store from '@ember-data/store';
 import Route from '@ember/routing/route';
-import RouterService from '@ember/routing/router-service';
-import { inject as service } from '@ember/service';
 
 
-export default class MetadataRoute extends Route {
-    @service store!: Store;
-    @service router!: RouterService;
-
-    async model() {
-        const params = this.modelFor('guid-node.metadata');
-
-        const cedarMetadataRecords = await params.guidNode.queryHasMany('cedarMetadataRecords', {
-            'page[size]': 20,
-        });
-
-        // This is for prototyping to get a working view to the mirage server
-        // This will be removed before production
-        // Brian - 2024-01-09
-        for(const item of cedarMetadataRecords) {
-            await item.template;
-        }
-
-        return {
-            guidNode: params.guidNode,
-            cedarMetadataRecords,
-        };
+export default class MetadataIndexRoute extends Route {
+    model() {
+        this.replaceWith('guid-node.metadata.detail', 'osf');
     }
 }
