+++ app/router.ts
@@ -73,8 +73,8 @@ Router.map(function() {
             this.route('provider', { path: '/:providerId' });
         });
         this.route('metadata', function() {
-            this.route('index', {path: '/:record_id'});
-            this.route('index', {path: '/'});
+            this.route('index', { path: '/'});
+            this.route('detail', { path: '/:record_id' });
             this.route('add');
         });
         this.route('registrations');
