+++ lib/osf-components/addon/components/cedar-metadata-editor/component.ts
@@ -43,7 +43,7 @@ export default class CedarMetadataEditor extends Component<Args> {
         record.metadata = cee.currentMetadata;
         await record.save();
         // eslint-disable-next-line no-debugger
-        debugger;
-        this.router.transitionTo('guid-node.metadata', this.args.target.id, record.id);
+        // debugger;
+        this.router.transitionTo('guid-node.metadata.detail', record.id);
     }
 }
